pyproject.toml
[project]
name = "content-pipeline"
version = "0.1.0"
description = "Синяя шляпа — automated cybersecurity content pipeline"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.28",
    "beautifulsoup4>=4.12",
    "python-telegram-bot>=22.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "anthropic>=0.86",
    "arxiv>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
